Introduce interface for registering remotes.
This abstracts how we reference RMI remotes, whether in the proxy or
by samplers.

In the case of the capture side (the RMI proxy), the proxy itself is
an instance registry, where later on we wire up the proxy invocation
handler to call into it when it sees a java.rmi.Remote instance being
returned to the client. Meanwhile, on the playback side, the RMI
remote object config contains an instance registry that is bound to
the thread test context.

Note that we've changed RMIRemoteObjectConfig to initialize the
registry and bind the registry to the context variables on test start,
and clear the registry on test end.