RMI Sampler: Bind BeanShell interpreter on test start.
Because JMeter serializes all bean properties to the JMX test plan,
the BeanShell interpreter ends up also being serialized, dumping large
amounts of interpreter state if the test plan is saved after a test
run. This is bad: we don't need to have interpreter state persisted
between runs, hence us marking the property as temporary. Apparently,
this is not enough.

Since it's fine for us to have a single interpreter per test thread,
just create the interpreter and do the initial eval of the argument
script at test start, and remove the interpreter at test end. Also,
make the getter for the interpreter private so it isn't visible to the
serialization code.