Make scriptlet generator instance-based.
The existing design was badly broken by our introduction of the cycle
detector (i.e. the state variable generatedObjects), because it would
now be possible to break a later call to generateScriptletForObject on
an instance that was previously generated in a previous call (e.g. to
serialize an argument on a previous RMI call), as the later call in
one context would then have an invalid reference to the previous
object in a different context.

So, change the scriptlet generator to be instance-based, and create a
new generator for every context/method call to generate arguments for,
so each context has a fresh state.